configure second lcd screen, and the inferno logo on it.
the second screen is also a frame-buffer like device:  16-bit pixels
(5 bits for each color of rgb, and a transparency (which we don't
use).  the framebuffer starts at 0x6200000.  it is 256x256 pixels.
the screen is 192 rows high, so we don't use the lower rows (the
visible part can be changed by setting a register).

the hard part was finding the erroneous defines for the bits that
enable power on the 2d subsystem for the second sceren, in lcdreg.h.

right now, the inferno logo is built-in, in file logo.c.  we probably
won't need it for long, as we want to display something more useful
on that secondary screen. :)  i tried reading the image file from
#/, but that failed with locking (i tried it with k(open|read|etc),
before schedinit(), that's probably not allowed).  the image file
is currently only ~8kb, due to using a simple run-length encoding
scheme (1 byte count, 2 bytes pixel).